+++ lib/ndr_dev_support/rake_ci/commit_cop.rb
@@ -2,6 +2,7 @@ require 'active_support/core_ext/module/attribute_accessors'
 require_relative 'commit_cop/migration_without_structure_dump'
 require_relative 'commit_cop/modified_migration'
 require_relative 'commit_cop/renamed_migration'
+require_relative 'commit_cop/missing_associated_test_file'
 
 module NdrDevSupport
   module RakeCI
@@ -16,12 +17,26 @@ module NdrDevSupport
           []
         end
 
+      # This defines the regular expression that identifies the path to application files
+      # that you would expect to have associated test files.
+      mattr_accessor :tested_paths
+      self.tested_paths =
+        if defined?(Rails)
+          Rails.application.config.paths.values_at('app/controllers',
+                                                   'app/helpers',
+                                                   'app/models',
+                                                   'app/mailers').flat_map { |obj| obj.map(&:to_s) }
+        else
+          []
+        end
+
       # This defines the regular expression that identifies structure dump files.
       mattr_accessor :structure_dump_pattern
       self.structure_dump_pattern = %r{\Adb/(structure\.sql|schema\.rb)\z}
 
       COMMIT_COPS = [
         MigrationWithoutStructureDump,
+        MissingAssociatedTestFile,
         ModifiedMigration,
         RenamedMigration
       ].freeze