+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Changed
 * Setting the Capybara save path and prune strategy. Resolves #45 (#46)
 
+### Added
+* Added an additional cop to check commits for missing associated test files.
+
 ## 4.2.1 / 2018-12-18
 ### Fixed
 * ci_server: improve logging and error handling (#44)